✨ Add custom server https://nextjs.org/docs/advanced-features/custom-server https://github.com/vercel/next.js/tree/canary/examples/custom-server-typescript
A custom server supports consistent connections with web sockets or databases and allows custom routing

Steps:

Add server folder with server/index.ts. This is the server entry point with a custom route for /storybook to allow a deployment of the app and storybook on a single server. All other requests are forwarded to next request handler.

- npm i -D nodemon ts-node which is required for the new dev script.
- Add tsconfig.server.json for a custom TypeScript config for the server side. The existing tsconfig is extended, but it's required to change the module
- Add nodemon.json to configure nodemon. The server folder is watched.
- Add and update dev, build-next, build-server, build and start scripts.
- npm run dev is used in development with hot module replacement.
npm run build creates optimized client, storybook and server builds.
npm start calls the server build.
- Please verify that all scripts are working and try out /storybook route.

Notes:
A custom server can not be deployed on Vercel, the platform Next.js was made for